+++ src/Serilog.Sinks.Udp/LoggerSinkConfigurationExtensions.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System;
+using System.ComponentModel;
 using System.Net;
 using Serilog.Configuration;
 using Serilog.Events;
@@ -30,36 +31,13 @@ namespace Serilog
         private const string DefaultOutputTemplate = "{Timestamp:yyyy-MM-dd HH:mm:ss} [{Level}] {Message}{NewLine}{Exception}";
 
         /// <summary>
-        /// Adds a sink that sends log events as UDP packages over the network.
+        /// Extension method providing JSON support via
+        /// <a href="https://github.com/serilog/serilog-settings-configuration">Serilog.Settings.Configuration</a>.
         /// </summary>
-        /// <param name="sinkConfiguration">Logger sink configuration.</param>
-        /// <param name="remoteAddressAsString">
-        /// The address of the remote host or multicast group to which the UDP
-        /// client should sent the logging event.
-        /// </param>
-        /// <param name="remotePort">
-        /// The TCP port of the remote host or multicast group to which the UDP client should sent
-        /// the logging event.
-        /// </param>
-        /// <param name="localPort">
-        /// The TCP port from which the UDP client will communicate. The default is 0 and will
-        /// cause the UDP client not to bind to a local port.
-        /// </param>
-        /// <param name="restrictedToMinimumLevel">
-        /// The minimum level for events passed through the sink. The default is
-        /// <see cref="LevelAlias.Minimum"/>.
-        /// </param>
-        /// <param name="outputTemplate">
-        /// A message template describing the format used to write to the sink. The default is
-        /// "{Timestamp:yyyy-MM-dd HH:mm:ss} [{Level}] {Message}{NewLine}{Exception}".
-        /// </param>
-        /// <param name="formatProvider">
-        /// Supplies culture-specific formatting information, or null.
-        /// </param>
-        /// <returns>Logger configuration, allowing configuration to continue.</returns>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public static LoggerConfiguration Udp(
             this LoggerSinkConfiguration sinkConfiguration,
-            string remoteAddressAsString,
+            string remoteAddress,
             int remotePort,
             int localPort = 0,
             LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum,
@@ -68,7 +46,7 @@ namespace Serilog
         {
             return Udp(
                 sinkConfiguration,
-                IPAddress.Parse(remoteAddressAsString),
+                IPAddress.Parse(remoteAddress),
                 remotePort,
                 localPort,
                 restrictedToMinimumLevel,
@@ -122,6 +100,28 @@ namespace Serilog
             return Udp(
                 sinkConfiguration,
                 remoteAddress,
+                remotePort,
+                formatter,
+                localPort,
+                restrictedToMinimumLevel);
+        }
+
+        /// <summary>
+        /// Extension method providing JSON support via
+        /// <a href="https://github.com/serilog/serilog-settings-configuration">Serilog.Settings.Configuration</a>.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public static LoggerConfiguration Udp(
+            this LoggerSinkConfiguration sinkConfiguration,
+            string remoteAddress,
+            int remotePort,
+            ITextFormatter formatter,
+            int localPort = 0,
+            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum)
+        {
+            return Udp(
+                sinkConfiguration,
+                IPAddress.Parse(remoteAddress),
                 remotePort,
                 formatter,
                 localPort,